feat: migrar archivos de autenticación y configuración de la tienda a TypeScript
