+++ client/src/App.tsx
@@ -1,7 +1,7 @@
 import './App.css'
 import MyRoutes from './routes/routes.jsx';
 import { Provider } from 'react-redux'
-import { store } from './app/store.js';
+import { store } from './app/store.ts';
 
 function App() {
   return (
