--- client/src/app/store.js
@@ -1,8 +1,0 @@
-import { configureStore } from "@reduxjs/toolkit";
-import authReducer from "../features/authSlice";
-
-export const store = configureStore({
-  reducer: {
-    authentication: authReducer,
-  },
-});
