+++ client/src/app/store.ts
@@ -1,0 +1,11 @@
+import { configureStore } from "@reduxjs/toolkit";
+import authReducer from "../features/authSlice.ts";
+
+export const store = configureStore({
+  reducer: {
+    authentication: authReducer,
+  },
+});
+
+export type RootState = ReturnType<typeof store.getState>;
+export type AppDispatch = typeof store.dispatch;
